% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amgut-data.R
\docType{data}
\name{amgut}
\alias{amgut}
\title{Microbiome family-level data from American Gut dataset}
\format{
List with two objects: \code{clin.dat} is a dataframe with 100 rows and 4 columns containing the covariate data; \code{counts} is a matrix with 100 rows and 10 columns which contains the family-level counts of each subject.  The last column of \code{otu.counts} is a reference category containing the sum of all families not included in this dataset.
}
\source{
\href{https://qiita.ucsd.edu/study/description/10317}{Qiita database}
}
\usage{
data(amgut)
}
\description{
Counts from 16S sequencing on the American Gut dataset.  Clinical covariates
also include
}
\examples{
data(amgut)

}
\references{
McDonald, Daniel, et al. "American gut: an open platform for citizen science microbiome research." MSystems 3.3 (2018): e00031-18.
}
\keyword{american}
\keyword{datasets}
\keyword{gut}
\keyword{microbiome}
