% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getPredCov}
\alias{getPredCov}
\title{Title}
\usage{
getPredCov(obj, newdata = NULL, quant = c(0.025, 0.975),
  type = c("cov", "cor", "prec", "pcor"), post.stat = c("mean",
  "median"))
}
\arguments{
\item{post.stat}{}
}
\description{
Title
}
