% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micore.R
\name{mc}
\alias{mc}
\title{Main function for running micore (not exported)}
\usage{
mc(
  counts,
  lr.counts,
  Xt,
  C0 = NULL,
  Psi0 = NULL,
  Gamma0 = NULL,
  nuPsi = NULL,
  nuGamma = NULL,
  target.accept.rate = 0.23,
  n.samp = 4000,
  n.burn = 4000,
  adapt.control = NULL,
  save.eta.cov = FALSE,
  verbose = FALSE
)
}
\description{
Main function for running micore (not exported)
}
