% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.micore}
\alias{print.micore}
\title{Print method for \code{micore} object.}
\usage{
\method{print}{micore}(x, ...)
}
\arguments{
\item{x}{An object of class \code{micore}}

\item{...}{Further arguments passed to or from other methods}
}
\value{
Prints information about \code{micore} object.
}
\description{
Print method for \code{micore} object.
}
