% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micore.R
\name{micore}
\alias{micore}
\title{Title}
\usage{
micore(counts, X, C0 = NULL, Psi0 = NULL, Gamma0 = NULL,
  nuPsi = NULL, nuGamma = NULL, n.chain = 4, n.cores = n.chain,
  target.accept.rate = 0.23, n.samp = 4000, n.burn = 4000,
  adapt.control = NULL, save.eta.cov = FALSE, verbose = FALSE)
}
\arguments{
\item{counts}{Matrix containing OTU counts.  Rows are samples, columns are OTUs.}

\item{X}{Design matrix contating covariates of interest. Rows are samples, columns are covariates.  Can include intercept column.}

\item{C0}{Prior mean for "C" parameter.  If null, then C0 is set automatically.}

\item{Psi0}{Prior inverse-Wishart matrix hyperparameter for "Psi" parameter.  If null, then Psi0 is set automatically.}

\item{Gamma0}{Prior inverse-Wishart matrix hyperparameter for "Gamma" parameter.  If null, then Gamma0 is set automatically.}

\item{nuPsi}{Prior inverse-Wishart degrees of freedom for "Psi" parameter.  If null, then nuPsi is set automatically.}

\item{nuGamma}{Prior inverse-Wishart degrees of freedom for "Gamma" parameter.  If null, then nuGamma is set automatically.}

\item{n.chain}{Number of independent MCMC chains to run in parallel}

\item{n.cores}{Number of cores to use to parallelize the MCMC chains}

\item{target.accept.rate}{Target acceptance rate for adaptive Metropolis sampler.}

\item{n.samp}{Number of MCMC samples}

\item{n.burn}{Number of burn-in samples}

\item{adapt.control}{List contatining tuning parameters for adaptive MCMC part.  See .}

\item{save.eta.cov}{Logical. Save the proposal distribution variance from the Metropolis sampler?  Caution, this will be large.}

\item{verbose}{Logical. Output progress?}
}
\description{
Title
}
