% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.micore}
\alias{predict.micore}
\title{Title}
\usage{
\method{predict}{micore}(
  object,
  newdata = NULL,
  type = c("alr", "proportion"),
  post.stat = c("mean", "median"),
  quant = c(0.025, 0.975),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
